SE-11: UI changes made.
