+++ client/src/components/Venues/ListVenues/ListVenues.scss
@@ -1,0 +1,10 @@
+.card .card-list img {
+  border-radius: 5px 5px 5px 5px;
+  width: 50%;
+}
+
+.card-list {
+  width: 90%;
+  margin-left: 5%;
+  margin-top: 1rem;
+}
