+++ client/src/components/Venues/VenueDetail/VenueDetail.scss
@@ -1,0 +1,9 @@
+.card .card-detail img {
+  border-radius: 5px 5px 5px 5px;
+  width: 50%;
+}
+
+.card-detail {
+  width: 55%;
+  margin-left: 17.5%;
+}
