+++ client/src/components/Venues/VenueEdit/VenueEdit.scss
@@ -1,0 +1,8 @@
+.btn-primary-edit {
+  margin-bottom: 2%;
+}
+
+.input-venue-edit {
+  width: 80% !important;
+  margin-left: 10%;
+}
