+++ client/src/components/auth/Register.js
@@ -39,69 +39,71 @@ const Register = ({ setAlert, register, isAuthenticated }) => {
 
   return (
     <Fragment>
-      <h1 className="large text-primary">Sign Up</h1>
-      <p className="lead">
-        <i className="fas fa-user"></i> Create Your Account
-      </p>
-      <form className="form" onSubmit={(e) => onSubmit(e)}>
-        <div className="form-group">
-          <p className="isAdmin">
-            <i className="fas fa-user-cog"></i> Is admin?
+      <section className="container">
+        <h1 className="large text-primary">Sign Up</h1>
+        <p className="lead">
+          <i className="fas fa-user"></i> Create Your Account
+        </p>
+        <form className="form" onSubmit={(e) => onSubmit(e)}>
+          <div className="form-group">
+            <p className="isAdmin">
+              <i className="fas fa-user-cog"></i> Is admin?
+              <input
+                type="checkbox"
+                id="isAdmin"
+                name="isAdmin"
+                onChange={(e) => onChange(e)}
+              ></input>
+            </p>
             <input
-              type="checkbox"
-              id="isAdmin"
-              name="isAdmin"
+              type="text"
+              placeholder="Name"
+              name="name"
+              value={name}
               onChange={(e) => onChange(e)}
-            ></input>
-          </p>
-          <input
-            type="text"
-            placeholder="Name"
-            name="name"
-            value={name}
-            onChange={(e) => onChange(e)}
-            required
-          />
-        </div>
-        <div className="form-group">
-          <input
-            type="email"
-            placeholder="Email Address"
-            name="email"
-            value={email}
-            onChange={(e) => onChange(e)}
-            required
-          />
-          <small className="form-text">
-            This site uses Gravatar so if you want a profile image, use a
-            Gravatar email
-          </small>
-        </div>
-        <div className="form-group">
-          <input
-            type="password"
-            placeholder="Password"
-            name="password"
-            value={password}
-            onChange={(e) => onChange(e)}
-            minLength="6"
-          />
-        </div>
-        <div className="form-group">
-          <input
-            type="password"
-            placeholder="Confirm Password"
-            name="password2"
-            value={password2}
-            onChange={(e) => onChange(e)}
-            minLength="6"
-          />
-        </div>
-        <input type="submit" className="btn btn-primary" value="Register" />
-      </form>
-      <p className="my-1">
-        Already have an account? <Link to="/login">Sign In</Link>
-      </p>
+              required
+            />
+          </div>
+          <div className="form-group">
+            <input
+              type="email"
+              placeholder="Email Address"
+              name="email"
+              value={email}
+              onChange={(e) => onChange(e)}
+              required
+            />
+            <small className="form-text">
+              This site uses Gravatar so if you want a profile image, use a
+              Gravatar email
+            </small>
+          </div>
+          <div className="form-group">
+            <input
+              type="password"
+              placeholder="Password"
+              name="password"
+              value={password}
+              onChange={(e) => onChange(e)}
+              minLength="6"
+            />
+          </div>
+          <div className="form-group">
+            <input
+              type="password"
+              placeholder="Confirm Password"
+              name="password2"
+              value={password2}
+              onChange={(e) => onChange(e)}
+              minLength="6"
+            />
+          </div>
+          <input type="submit" className="btn btn-primary" value="Register" />
+        </form>
+        <p className="my-1">
+          Already have an account? <Link to="/login">Sign In</Link>
+        </p>
+      </section>
     </Fragment>
   );
 };
