+++ client/src/components/layout/Alert/Alert.js
@@ -1,14 +1,19 @@
-import React from "react";
+import { React, Fragment } from "react";
 import PropTypes from "prop-types";
 import { connect } from "react-redux";
+import "./Alert.scss";
 
 const Alert = ({ alerts }) =>
   alerts !== null &&
   alerts.length > 0 &&
   alerts.map((alert) => (
-    <div key={alert.id} className={`alert alert-${alert.alertType}`}>
-      {alert.msg}
-    </div>
+    <Fragment>
+      {/* <section className="container"> */}
+      <div key={alert.id} className={`alert alert-${alert.alertType}`}>
+        {alert.msg}
+      </div>
+      {/* </section> */}
+    </Fragment>
   ));
 
 Alert.propTypes = {
