+++ client/src/components/layout/Landing/Landing.js
@@ -2,6 +2,7 @@ import React, { Fragment } from "react";
 import { Link } from "react-router-dom";
 import { connect } from "react-redux";
 import { PropTypes } from "prop-types";
+import "./Landing.scss";
 
 const Landing = ({ isAuthenticated }) => {
   const guestButtons = (
