+++ client/src/components/layout/Landing/Landing.scss
@@ -1,0 +1,20 @@
+/* Landing Page */
+
+.landing {
+  position: relative;
+  background: url("../../../img/sports-landing-bg.jpg") no-repeat center
+    (center / cover);
+  height: 100vh;
+}
+
+.landing-inner {
+  color: #fff;
+  height: 100%;
+  width: 80%;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: center;
+  text-align: center;
+}
