+++ client/src/components/layout/Navbar/Navbar.scss
@@ -1,0 +1,58 @@
+/* Navbar */
+
+.navbar {
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  padding: 0.7rem 2rem;
+  position: fixed;
+  z-index: 1;
+  width: 100%;
+  top: 0;
+  border-bottom: solid 1px var(--primary-color);
+  opacity: 0.9;
+
+  ul {
+    display: flex;
+  }
+
+  a {
+    color: #fff;
+    padding: 0.45rem;
+    margin: 0 0.25rem;
+
+    &:hover {
+      color: var(--primary-color);
+    }
+  }
+
+  h1 {
+    font-size: 1.5rem;
+  }
+
+  .welcome span {
+    margin-right: 0.6rem;
+  }
+}
+
+@media (max-width: 700px) {
+  /* Navbar */
+
+  .navbar {
+    display: block;
+    text-align: center;
+
+    ul {
+      text-align: center;
+      justify-content: center;
+    }
+
+    h1 {
+      margin-bottom: 1rem;
+    }
+
+    .welcome {
+      display: none;
+    }
+  }
+}
