+++ config/db.js
@@ -4,12 +4,12 @@ const db = config.get("mongoURI");
 
 const connectDB = async () => {
   try {
-    await mongoose.connect(db, {
-      useNewUrlParser: true,
-    });
-    // await mongoose.connect("mongodb://localhost:27017/SEdb", {
+    // await mongoose.connect(db, {
     //   useNewUrlParser: true,
     // });
+    await mongoose.connect("mongodb://localhost:27017/SEdb", {
+      useNewUrlParser: true,
+    });
     console.log("MongoDB connected");
   } catch (err) {
     console.error(err.message);
