+++ package.json
@@ -31,6 +31,7 @@
   },
   "devDependencies": {
     "concurrently": "^6.3.0",
-    "nodemon": "^2.0.13"
+    "nodemon": "^2.0.13",
+    "sass": "^1.49.0"
   }
 }
